+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.1.0] - 2025-06-01
+
+### Added
+
+- Method `clone` for creating copies of the static type manager configuration.
+- Method `clear` for removing all registrations from the type manager instance.
+
 ## [8.0.1] - 2025-03-22
 
 ### Changed